requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-braces==1.9.0
django-extra-views==0.8.0
django-guardian==1.4.4
django-postgrespool==0.3.0
django-registration-redux==1.4
django-s3direct==0.4.0
django-smart-selects==1.2.2
django-toolbelt==0.0.1
django-vanilla-views==1.0.4
django-widget-tweaks==1.4.1
gunicorn==19.6.0
Pillow==3.2.0
pluggy==0.3.1
psycopg2==2.6.1
py==1.4.31
six==1.10.0
SQLAlchemy==1.0.13
static3==0.7.0
tox==2.3.1
Unipath==1.1
virtualenv==15.0.2
whitenoise==3.2
